Move timezones and tests tasks into separate files
